Title: 10개 입력받기, Time: 151ms, Memory: 10MB, Status: Passed - CodeTree

+++ 240909/10개 입력받기/receive-10-inputs.java
@@ -22,7 +22,7 @@ public class Main {
             sum += n[i];
         }
         System.out.print(sum + " ");
-        avg = sum / index;
+        avg = sum / (double)index;
         System.out.printf("%.1f", avg);
     }
 }
